pyproject.toml
[project]
name = "yolo-labelstuio-kits"
version = "0.1.0"
description = "Full YOLO + Label Studio integration kit for semi-automatic dataset labeling"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "label-studio>=1.9.0",
    "label-studio-sdk>=0.0.32",
    "ultralytics>=8.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]
